+++ source/SoloReadFeature.h
@@ -35,8 +35,7 @@ public:
     void addCounts(const SoloReadFeature &soloCBin);
     void addStats(const SoloReadFeature &soloCBin);
     void statsOut(ofstream &streamOut);
-    void inputRecords(uint32 **cbP, uint32 cbPstride, uint32 *cbReadCountExact, 
-                      vector<readInfoStruct> &readInfo, SoloFeature **soloFeatAll);
+    void inputRecords(uint32 **cbP, uint32 cbPstride, uint32 *cbReadCountExact, vector<readInfoStruct> &readInfo, SoloFeature *soloFeat);
 
 private:
     const int32 featureType;